# docs/swagger.yaml
definitions:
  auth.CreateUserRequest:
    properties:
      account:
        type: string
      account_password:
        type: string
      avatar:
        type: string
      nickname:
        type: string
    type: object
...
  collection.CreateCollectionRequest:
    properties:
      idea_id:
        type: integer
    type: object
  collection.DeleteCollectionRequest:
    properties:
      idea_id:
        type: integer
    type: object
  collection.IdeaResponse:
    properties:
      count:
        type: integer
      list:
        items:
          $ref: '#/definitions/model.IdeaListItem'
        type: array
    type: object
  errno.Errno:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  idea.CommentResponse:
    properties:
      count:
        type: integer
      list:
        items:
          $ref: '#/definitions/model.CommentListItem'
        type: array
    type: object
  idea.CreateCommentRequest:
    properties:
      commented_id:
        type: integer
      content:
        type: string
    type: object
...
  model.CommentListItem:
    properties:
      avatar:
        type: string
      commenter_id:
        type: integer
      content:
        type: string
      id:
        type: integer
      idea_id:
        type: integer
      liked:
        description: 是否点赞
        type: boolean
      likes_sum:
        type: integer
      nickname:
        type: string
      release_date:
        type: string
    type: object
...
  model.IdeaListItem:
    properties:
      avatar:
        type: string
      comment_sum:
        type: integer
      content:
        type: string
      idea_id:
        type: integer
      liked:
        description: 是否点赞
        type: boolean
      likes_sum:
        type: integer
      nickname:
        type: string
      publisher_id:
        type: integer
      release_date:
        type: string
    type: object
...
host: 124.71.184.107
info:
  contact:
    email: 2036479155@qq.com
    name: gongna
  description: 零重力APP
  termsOfService: http://swagger.io/terrms
  title: ZeroGravity
  version: 1.0.0
paths:
...
  /api/v1/auth/api/v1/register:
    post:
      consumes:
      - application/json
      description: 用户注册
      parameters:
      - description: "Account--账户 AccountPassword--账户密码\tNickName--昵称\tAvatar--头像 "
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/auth.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errno.Errno'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errno.Errno'
      summary: 用户输入账号和密码注册
      tags:
      - auth
  /api/v1/collection:
    delete:
      consumes:
      - application/json
      description: 删除收藏记录
      parameters:
      - description: userid--用户的ID
        in: header
        name: token
        required: true
        type: string
      - description: IdeaId用户收藏的想法ID
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/collection.DeleteCollectionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errno.Errno'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errno.Errno'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errno.Errno'
      summary: 用户删除收藏记录
      tags:
      - collection
    get:
      consumes:
      - application/json
      description: 查看用户的收藏列表
      parameters:
      - description: userid用户ID
        in: header
        name: token
        required: true
        type: string
      - description: 'limit--偏移量指定开始返回记录之前要跳过的记录数 '
        in: query
        name: limit
        required: true
        type: integer
      - description: 'page--限制指定要检索的记录数 '
        in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":0,"message":"OK","data":{"username":"kong"}}'
          schema:
            $ref: '#/definitions/collection.IdeaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errno.Errno'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errno.Errno'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errno.Errno'
      summary: 查询用户的收藏列表
      tags:
      - collection
    post:
      consumes:
      - application/json
      description: 新增收藏记录
      parameters:
      - description: userid--用户的ID
        in: header
        name: token
        required: true
        type: string
      - description: IdeaId用户收藏的想法ID
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/collection.CreateCollectionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errno.Errno'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errno.Errno'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errno.Errno'
      summary: 添加收藏记录
      tags:
      - collection
...
  /api/v1/idea/detail/:id:
    delete:
      consumes:
      - application/json
      description: 删除想法
      parameters:
      - description: userid--用户ID
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errno.Errno'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errno.Errno'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errno.Errno'
      summary: 删除想法
      tags:
      - idea
    get:
      consumes:
      - application/json
      description: 根据用户ID获取这个ID下发布的想法
      parameters:
      - description: id--根据这个id判断发布的想法是否为私密，id为1是显示想法为私密发布，你无权查看
        in: path
        name: id
        required: true
        type: integer
      - description: uid--用户ID
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":0,"message":"OK","data":{"username":"kong"}}'
          schema:
            $ref: '#/definitions/model.IdeaListItem'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errno.Errno'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errno.Errno'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errno.Errno'
      summary: 获取想法
      tags:
      - idea
  /api/v1/idea/detail/:id/comment:
    get:
      consumes:
      - application/json
      description: 想法的评论列表
      parameters:
      - description: uid--用户ID
        in: header
        name: token
        required: true
        type: string
      - description: id--被评论的想法ID
        in: path
        name: id
        required: true
        type: integer
      - description: limit--偏移量指定开始返回记录之前要跳过的记录数
        in: query
        name: limit
        required: true
        type: integer
      - description: 'page--限制指定要检索的记录数 '
        in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":0,"message":"OK","data":{"username":"kong"}}'
          schema:
            $ref: '#/definitions/idea.CommentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errno.Errno'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errno.Errno'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errno.Errno'
      summary: 发布的想法下对应的评论
      tags:
      - idea
    post:
      consumes:
      - application/json
      description: Add comment records to the database
      parameters:
      - description: userid--用户的ID
        in: header
        name: token
        required: true
        type: string
      - description: CommentedId 被评论者的ID || Content 评论内容
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/idea.CreateCommentRequest'
      - description: IdeaId--想法ID
        in: path
        name: idea_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errno.Errno'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errno.Errno'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errno.Errno'
      summary: Add comment
      tags:
      - idea
  /api/v1/idea/detail/:id/comment/:comment_id:
    delete:
      consumes:
      - application/json
      description: 删除评论
      parameters:
      - description: id--评论的ID
        in: path
        name: id
        required: true
        type: string
      - description: uid --用户的ID
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errno.Errno'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errno.Errno'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errno.Errno'
      summary: 删除用户自己的评论
      tags:
      - idea
...
schemes:
- http
swagger: "2.0"